+++ src/qnethernet/drivers/driver_teensy41.c
@@ -37,6 +37,9 @@
 //  Defines
 // --------------------------------------------------------------------------
 
+#define MTU           1500
+#define MAX_FRAME_LEN 1518
+
 static const uint32_t GPIO_PAD_OUTPUT = (0
     /* HYS_0_Hysteresis_Disabled */
     /* PUS_0_100K_Ohm_Pull_Down */
@@ -823,6 +826,14 @@ FLASHMEM void driver_get_capabilities(struct DriverCapabilities* const dc) {
   dc->isAutoNegotiationSettable    = true;
   dc->isLinkCrossoverDetectable    = true;
   dc->isAutoNegotiationRestartable = true;
+}
+
+size_t driver_get_mtu(void) {
+  return MTU;
+}
+
+size_t driver_get_max_frame_len(void) {
+  return MAX_FRAME_LEN;
 }
 
 bool driver_is_unknown(void) {